Add new UI container and text components
This commit adds a new Card container and text components (like Heading and Paragraph) to the UI library. Corresponding styles for each new component are also added along with the test cases in 'main.ts' to understand the component's functionality.
This would widen the variety of containers and text components offered by the UI library for the users who can choose from these according to their design and layout needs.

Logic and style requirements for new components are defined in respective '.ts' and '.scss' files. The added UI components also have been categorized and exported in 'components.ts' file to ease the import. For debugging purposes, a new 'dumpComponents' method is added which returns an array of all currently-loaded UI components.

The main reason for this commit is to enhance the UI library by incorporating commonly used components with flexibility for alterations according to user needs.